blobs: Update linuxkit binaries to version based on v1.2.0
The linuxkit blobs provided as part of this repository have been a sore
point with regards to how these are built and where they come from. This
updates them to a new version that is based on top of the latest
upstream release v1.2.0.

With the update linuxkit has changed the format of its command line
options from the -<option> format to the --<option> format.
Further, linuxkit introduced functionality to add an SBOM file with
non-deterministic unique IDs, which breaks our image reproducibility.
Pass `--no-sbom` to disable that SBOM creation.

The patches we carry on top of that upstream release are now available
through aws-nitro-enclaves-sdk-bootstrap package, which now also
provides a way to reproducibly build these binaries using nix.

With the update we have cleaned up our downstream patches, which entail
some changes on how we invoke linuxkit:

1) The downstream `prefix` functionality has been moved into the
   configuration yaml files instead of having it as a CLI option.
2) We use a new output type `kernel+initrd-nogz` which produces the
   initrd with cpio instead of the standard cpio.gz.

As we have always had functionally similar patches on top of upstream
linuxkit there are no fcuntional changes.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>