build-enclave: Implement arguments to override build time
There may be cases where overriding the build time for an EIF in the
metadata is beneficial (e.g. reproducible builds). Provide the necessary
implementation to set a built time override and instruct the build info
generation to act accordingly.

If a date is passed that can not be parsed according to RFC3339, we
choose UNIX_EPOCH (1970-01-01 00:00:00 UTC).

Signed-off-by: Leonard Foerster <foersleo@amazon.de>